+++ com.unity.netcode.adapter.utp/Runtime/UnityTransport.cs
@@ -170,8 +170,6 @@ namespace Unity.Netcode
         private State m_State = State.Disconnected;
         private NetworkDriver m_Driver;
         private NetworkSettings m_NetworkSettings;
-        private byte[] m_MessageBuffer;
-        private NetworkConnection m_ServerConnection;
         private ulong m_ServerClientId;
 
         private NetworkPipeline m_UnreliableFragmentedPipeline;
@@ -183,6 +181,8 @@ namespace Unity.Netcode
         public ProtocolType Protocol => m_ProtocolType;
 
         private RelayServerData m_RelayServerData;
+
+        internal NetworkManager NetworkManager;
 
 #if UNITY_EDITOR
         private static int ClientPacketDelayMs => UnityEditor.EditorPrefs.GetInt($"NetcodeGameObjects_{Application.productName}_ClientDelay");
@@ -309,8 +309,8 @@ namespace Unity.Netcode
                 return false;
             }
 
-            m_ServerConnection = m_Driver.Connect(serverEndpoint);
-            m_ServerClientId = ParseClientId(m_ServerConnection);
+            var serverConnection = m_Driver.Connect(serverEndpoint);
+            m_ServerClientId = ParseClientId(serverConnection);
 
             return true;
         }
@@ -530,11 +530,14 @@ namespace Unity.Netcode
                 {
                     // Some error occured. If it's just the UTP queue being full, then don't log
                     // anything since that's okay (the unsent message(s) are still in the queue
-                    // and we'll retry sending the later);
+                    // and we'll retry sending the later). Otherwise log the error and remove the
+                    // message from the queue (we don't want to resend it again since we'll likely
+                    // just get the same error again).
                     if (result != (int)Networking.Transport.Error.StatusCode.NetworkSendQueueFull)
                     {
                         Debug.LogError("Error sending the message: " +
                             ErrorUtilities.ErrorToString((Networking.Transport.Error.StatusCode)result, clientId));
+                        queue.Consume(written);
                     }
 
                     return;
@@ -596,13 +599,14 @@ namespace Unity.Netcode
         private bool ProcessEvent()
         {
             var eventType = m_Driver.PopEvent(out var networkConnection, out var reader, out var pipeline);
+            var clientId = ParseClientId(networkConnection);
 
             switch (eventType)
             {
                 case TransportNetworkEvent.Type.Connect:
                     {
                         InvokeOnTransportEvent(NetcodeNetworkEvent.Connect,
-                            ParseClientId(networkConnection),
+                            clientId,
                             default(ArraySegment<byte>),
                             Time.realtimeSinceStartup);
 
@@ -611,30 +615,34 @@ namespace Unity.Netcode
                     }
                 case TransportNetworkEvent.Type.Disconnect:
                     {
-                        if (m_ServerConnection.IsCreated)
+                        // Handle cases where we're a client receiving a Disconnect event. The
+                        // meaning of the event depends on our current state. If we were connected
+                        // then it means we got disconnected. If we were disconnected means that our
+                        // connection attempt has failed.
+                        if (m_State == State.Connected)
                         {
-                            m_ServerConnection = default;
-
-                            var reason = reader.ReadByte();
-                            if (reason == (byte)Networking.Transport.Error.DisconnectReason.MaxConnectionAttempts)
-                            {
-                                Debug.LogError("Client failed to connect to server");
-                            }
+                            m_State = State.Disconnected;
+                            m_ServerClientId = default;
+                        }
+                        else if (m_State == State.Disconnected)
+                        {
+                            Debug.LogError("Failed to connect to server.");
+                            m_ServerClientId = default;
                         }
 
-                        m_ReliableReceiveQueues.Remove(ParseClientId(networkConnection));
+                        m_ReliableReceiveQueues.Remove(clientId);
+                        ClearSendQueuesForClientId(clientId);
 
                         InvokeOnTransportEvent(NetcodeNetworkEvent.Disconnect,
-                            ParseClientId(networkConnection),
+                            clientId,
                             default(ArraySegment<byte>),
                             Time.realtimeSinceStartup);
 
-                        m_State = State.Disconnected;
                         return true;
                     }
                 case TransportNetworkEvent.Type.Data:
                     {
-                        ReceiveMessages(ParseClientId(networkConnection), pipeline, reader);
+                        ReceiveMessages(clientId, pipeline, reader);
                         return true;
                     }
             }
@@ -707,6 +715,7 @@ namespace Unity.Netcode
                 {
                     m_State = State.Disconnected;
 
+                    m_ReliableReceiveQueues.Remove(m_ServerClientId);
                     ClearSendQueuesForClientId(m_ServerClientId);
 
                     // If we successfully disconnect we dispatch a local disconnect message
@@ -733,6 +742,7 @@ namespace Unity.Netcode
                     m_Driver.Disconnect(connection);
                 }
 
+                m_ReliableReceiveQueues.Remove(clientId);
                 ClearSendQueuesForClientId(clientId);
             }
         }
@@ -742,10 +752,12 @@ namespace Unity.Netcode
             return 0;
         }
 
-        public override void Initialize()
+        public override void Initialize(NetworkManager networkManager = null)
         {
             Debug.Assert(sizeof(ulong) == UnsafeUtility.SizeOf<NetworkConnection>(),
                 "Netcode connection id size does not match UTP connection id size");
+
+            NetworkManager = networkManager;
 
             m_NetworkSettings = new NetworkSettings(Allocator.Persistent);